chore(rename): Renamed tfm oss commands to tfm core
