+++ cmd/core/core.go
@@ -1,0 +1,19 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
+package core
+
+import (
+	"github.com/spf13/cobra"
+)
+
+// `tfm core` commands
+var coreCmd = &cobra.Command{
+	Use:   "core",
+	Short: "Command used to perform terraform open source (core) to TFE/TFC migration commands",
+	Long:  "Command used to perform terraform open source (core) to TFE/TFC migration commands",
+}
+
+func init() {
+
+}
